Commit4
Changed index page to bar graph for first file. Working on rendering the SVG properly
