--- SVGCodingTest/Pages/SVGQuantChange.cshtml
@@ -1,19 +1,0 @@
-﻿@page
-@model SVGCodingTest.Pages.SVGQuantChangeModel
-@{
-    ViewData["Title"] = "SVGQuantChange";
-}
-
-<h2>SVGQuantChange</h2>
-<head>
-    <script type="text/javascript">
-        
-    </script>
-</head>
-<body>
-    <div id="chartcontainer">
-        <svg id="svgcontainer" height="600" width="550">
-        </svg>
-    </div>
-</body>
-
