--- SVGCodingTest/Pages/SVGQuantChange.cshtml.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace SVGCodingTest.Pages
-{
-    public class SVGQuantChangeModel : PageModel
-    {
-        public List<Vehicle> vehicles = new List<Vehicle>();
-        VehicleMethods vehicleMethods = new VehicleMethods();
-        private List<Vehicle> PopulateVehicles()
-        {
-            vehicles.AddRange(new List<Vehicle>
-            {
-                new Vehicle("Bicycle", 200, 220),
-                new Vehicle("Car", 100, 94),
-                new Vehicle("Truck", 150, 168),
-                new Vehicle("Motorcycle", 40, 55),
-                new Vehicle("Jet", 11, 14)
-            });
-
-            vehicles = vehicleMethods.MarketShares(vehicles);
-
-            return vehicles;
-        }
-
-        public void OnGet()
-        {
-            vehicles = PopulateVehicles();
-        }
-    }
-}
