+++ SVGCodingTest/Vehicle.cs
@@ -7,7 +7,7 @@ namespace SVGCodingTest
 {
     public class Vehicle
     {
-        public Vehicle(string type, int y1, int y2) => (Type, QuantSoldY1, QuantSoldY2, MarketY1, MarketY2, Diff, TopYear) = (type, y1, y2, 0, 0, QuantDiff(), TopYearCalc());
+        public Vehicle(string type, int y1, int y2) => (Type, QuantSoldY1, QuantSoldY2, MarketY1, MarketY2, Diff) = (type, y1, y2, 0, 0, QuantDiff());
         public string Type { get; set; }
 
         public int QuantSoldY1 { get; set; }
@@ -20,17 +20,9 @@ namespace SVGCodingTest
 
         public int Diff { get; set; }
 
-        public int TopYear { get; set; }
-
-        private int TopYearCalc()
-        {
-            var year = QuantSoldY1 > QuantSoldY2 ? 1 : 2;
-            return year;
-        }
-
         private int QuantDiff()
         {
-            var result = QuantSoldY1 > QuantSoldY2 ? QuantSoldY1 - QuantSoldY2 : QuantSoldY2 - QuantSoldY1;
+            var result = QuantSoldY1 - QuantSoldY2;
             return result;
         }
     }
